directory restructure and rooting functionality
