+++ src/App.js
@@ -2,12 +2,12 @@ import React, { useState } from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Header from './components/Header.js';
 // import Footer from "./components/Footer";
-// import About from "./pages/About";
-// import Skills from "./pages/Skills";
-// import Content from "./pages/Content";
-// import Contact from "./pages/Contact";
-import SplashScreen from './SplashScreen';
-import './App.css';
+import Content from "./pages/Content.js";
+import MoreContent from "./pages/MoreContent.js";
+import About from "./pages/About.js";
+import Contact from "./pages/Contact.js";
+import SplashScreen from './pages/SplashScreen.js';
+import './styles/App.css';
 
 const App = () => {
   const [isLoading, setIsLoading] = useState(true); // Set this flag to indicate if the app is still loading or ready
@@ -26,11 +26,10 @@ const App = () => {
           <Header />
           <Routes>
             <Route path="/" element={<SplashScreen />} />
-            {/* Add other routes for your template pages */}
-            {/* <Route path="/About" element={<About />} />
+            <Route path="/Content" element={<Content />} /> 
             <Route path="/MoreContent" element={<MoreContent />} />
-            <Route path="/Content" element={<Content />} />
-            <Route path="/Contact" element={<Contact />} /> */}
+            <Route path="/About" element={<About />} />
+            <Route path="/Contact" element={<Contact />} />
           </Routes>
           {/* <Footer /> */}
         </Router>
