+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import App from './App';
 
 createRoot(document.getElementById('root')).render(<App />);
