+++ src/pages/MoreContent.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const MoreContent = () => {
+  return (
+    <div>
+      <h2>More Content Page</h2>
+      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed at mi nec purus tempor fringilla sed id lectus. Vivamus id mauris euismod, posuere felis sed, malesuada urna. Aliquam suscipit efficitur arcu, vel eleifend urna consequat a.</p>
+    </div>
+  );
+};
+
+export default MoreContent;
