+++ src/styles/SplashScreen.css
@@ -1,6 +1,6 @@
 .splash-screen {
     background-color: #293250;
-    background-image: url('splash.webp');
+    background-image: url('../splash.webp');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
